Matrix determinant/inverse; use double for elapsed time

+++ src/ta_timer.h
@@ -3,6 +3,6 @@
 
 void ta_timer_init();
 u64 ta_timer_elapsed_ticks();
-u64 ta_timer_elapsed_ms();
+double ta_timer_elapsed_ms();
 double ta_timer_elapsed_sec();
 u64 ta_timer_only_ms();
